refactor(rust): split convert and helpers into focused modules
Break the 2347-line convert.rs into a convert/ module dir (mod, output,
parse, plugins) and dissolve the 498-line helpers.rs grab-bag into
concept-named modules: url, scan, selector, tailwind. The merged entity
decoder moves into entities.rs alongside its lookup table.

Pure code movement: release builds use lto + codegen-units=1 so module
boundaries are invisible to codegen, and output is byte-identical.
Adds 22 unit tests for the newly isolated pure functions.
